pyproject.toml
[tool.poetry]
name = "tracker-of-habits"
version = "0.1.0"
description = ""
authors = ["Arthur Gaetov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"


[tool.poetry.group.dev.dependencies]
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
black = "^24.4.2"
djangorestframework-simplejwt = "^5.3.1"
django-filer = "^3.1.3"
django-filter = "^24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
